+++ index.html
@@ -12,7 +12,6 @@
 <body>
   <main>
     <script src="\js\functions.js"></script>
-    <script src="\js\main.js"></script>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
@@ -232,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
